+++ NEWS.md
@@ -1,5 +1,32 @@
-piecepackr 1.4.2
+piecepackr 1.5.0
 ================
+
+Breaking changes
+----------------
+
+There are no user-facing breaking changes but the internal `grid` structure of the grobs drawn/returned by `grid.piece()` / `pieceGrob()`
+have been changed to make it easier to query or edit the grid graphics post initial drawing
+with functions like `grid::grid.get()`, `grid::grid.edit()`, `grid::grid.reorder()` etc. (#205).
+
+* The `gTree` "grob" returned by `grid.piece()` and `pieceGrob()` now have a "piece" class and "named slots" which match the arguments of `grid.piece()`.
+* The new "piece" class grob only generates its content to draw at drawing time, one must use `grid::grid.force()` to view/edit its internal grob "children".
+* The internal grob layout for included grob functions are better named to more easily understand the structure and generate a matching ``grid::gPath``.
+
+Deprecated functions
+--------------------
+
+The following utility functions have been deprecated in favor of methods provided by the new `pp_shape()` object:
+
+* `get_shape_grob_fn()`, `pp_shape()$shape` instead.
+* `checkersGrob()`, use `pp_shape()$checkers()` instead.
+* `gridlinesGrob()`, use `pp_shape()$gridlines()` instead.
+* `hexlinesGrob()`, use `pp_shape()$hexlines()` instead.
+* `matGrob()`, use `pp_shape()$mat()` instead.
+
+New features
+------------
+
+* `pp_shape()` returns an R6 object with methods to create various grobs (`shape()`, `mat()`, `gridlines()`, `checkers()`, `hexlines()`).
 
 Bug fixes and minor improvements
 --------------------------------
@@ -11,6 +38,9 @@ Bug fixes and minor improvements
   with non-unit axis vectors and/or negative `axis_z` values.
 * Improved drawing of oblique projection edges for the "oval" shape (#212) and the "halma" shape.
 * `is_color_invisible()` will now correctly classify as "invisible" colors with an alpha channel value set to 0.
+* `pp_cfg()` now supports a "roundrect" (rounded rectangle) `shape` (#214).  Curvature of the corners are controlled by the `shape_r` style.
+* The named list returned `pp_cfg()$get_piece_opt()` now also contains a `back` value indicating whether this is a back of a piece
+  (and hence for some shapes may need to be flipped across a vertical line).
 
 piecepackr 1.4.1
 ================